devtools: Move `debugger_value_to_json` to `lib.rs`
