+++ components/devtools/lib.rs
@@ -35,6 +35,7 @@ use rand::{RngCore, rng};
 use resource::{ResourceArrayType, ResourceAvailable};
 use rustc_hash::FxHashMap;
 use serde::Serialize;
+use serde_json::{Map, Number, Value};
 use servo_base::generic_channel::{self, GenericSender};
 use servo_base::id::{BrowsingContextId, PipelineId, WebViewId};
 use servo_config::pref;
@@ -48,6 +49,7 @@ use crate::actors::framerate::FramerateActor;
 use crate::actors::inspector::InspectorActor;
 use crate::actors::inspector::walker::WalkerActor;
 use crate::actors::network_event::NetworkEventActor;
+use crate::actors::object::ObjectActor;
 use crate::actors::pause::PauseActor;
 use crate::actors::root::RootActor;
 use crate::actors::source::SourceActor;
@@ -946,4 +948,50 @@ fn handle_client(
     let _ = sender.send(DevtoolsControlMsg::ClientExited);
 
     registry.cleanup(stream_id);
+}
+
+/// <https://searchfox.org/mozilla-central/source/devtools/server/actors/object/utils.js#148>
+pub(crate) fn debugger_value_to_json(registry: &ActorRegistry, value: DebuggerValue) -> Value {
+    let mut v = Map::new();
+    match value {
+        DebuggerValue::VoidValue => {
+            v.insert("type".to_owned(), Value::String("undefined".to_owned()));
+            Value::Object(v)
+        },
+        DebuggerValue::NullValue => {
+            v.insert("type".to_owned(), Value::String("null".to_owned()));
+            Value::Object(v)
+        },
+        DebuggerValue::BooleanValue(boolean) => Value::Bool(boolean),
+        DebuggerValue::NumberValue(val) => {
+            if val.is_nan() {
+                v.insert("type".to_owned(), Value::String("NaN".to_owned()));
+                Value::Object(v)
+            } else if val.is_infinite() {
+                if val < 0. {
+                    v.insert("type".to_owned(), Value::String("-Infinity".to_owned()));
+                } else {
+                    v.insert("type".to_owned(), Value::String("Infinity".to_owned()));
+                }
+                Value::Object(v)
+            } else if val == 0. && val.is_sign_negative() {
+                v.insert("type".to_owned(), Value::String("-0".to_owned()));
+                Value::Object(v)
+            } else {
+                Value::Number(Number::from_f64(val).unwrap())
+            }
+        },
+        DebuggerValue::StringValue(str) => Value::String(str),
+        DebuggerValue::ObjectValue {
+            uuid,
+            class,
+            preview,
+            ..
+        } => {
+            let object_name = ObjectActor::register(registry, Some(uuid), class, preview);
+            let object_msg = registry.encode::<ObjectActor, _>(&object_name);
+            let value = serde_json::to_value(object_msg).unwrap_or_default();
+            Value::Object(value.as_object().cloned().unwrap_or_default())
+        },
+    }
 }
